pyproject.toml
[project]
name = "ai-test-generation-agent"
version = "0.1.0"
description = "Skeleton for evaluating AI-driven test generation with Gemini."
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.0",
    "pytest>=8.0.0",
    "coverage>=7.0.0",
    "mutmut>=2.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
branch = true
